requirements.txt
pyppeteer==1.0.2
python-dotenv==1.0.0
Pillow>=9.0.0

# 验证码识别依赖
ddddocr>=1.4.11
requests==2.31.0
beautifulsoup4==4.12.2 

# 定时调度
schedule==1.2.0 